@@ -1,12 +1,7 @@
 ﻿using System.ComponentModel;
-using System.Threading.Tasks;
 
-namespace CallingApp.Core
+namespace CallingApp.Core.ViewModels;
+
+public interface IBasePageViewModel : INotifyPropertyChanged
 {
-    public interface IBasePageViewModel : INotifyPropertyChanged
-    {
-        Task OnPageDisappearing();
-        Task OnPageAppearing();
-        Task OnPageCreated(params object[] parameters);
-    }
 }
